Misc fixes - dataset generation and episode sampling logic, training code, mps device option for mac (#36)
* Fix bugs in dataset generation

* update sampling logic and implement option for mps device on macs

* last fix on multi-config dataset generation

* Update offset for sampling eps for v short eps

* sampling episodes (again)

+++ src/agent/agent.py
@@ -53,4 +53,7 @@ class RandomAgent(Agent):
         return torch.tensor(0.0, requires_grad=True)
 
     def get_action(self, input: AgentInput, context=1, one_hot=False):
-        return torch.tensor(np.random.random(self.act_dim), device=input.states.device)
+        return torch.tensor(
+            np.random.random(self.act_dim).astype(np.float32),
+            device=input.states.device,
+        )
